delegate logic from payment prompt dialog
